# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Timewise database management system",
        "title": "timewise-dbms",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/dbms/v1/board_columns": {
            "post": {
                "description": "Create board column",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Create board column",
                "parameters": [
                    {
                        "description": "Create board column request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board_columns_dtos.BoardColumnsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                }
            }
        },
        "/dbms/v1/board_columns/rage/position": {
            "get": {
                "description": "Get rage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Get rage",
                "parameters": [
                    {
                        "description": "Get rage request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board_columns.RageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TwBoardColumn"
                            }
                        }
                    }
                }
            }
        },
        "/dbms/v1/board_columns/update_position/position": {
            "put": {
                "description": "Update position",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Update position",
                "parameters": [
                    {
                        "description": "Update position request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/dbms/v1/board_columns/update_position_after_deletion": {
            "put": {
                "description": "Update position after deletion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Update position after deletion",
                "parameters": [
                    {
                        "description": "Update position after deletion request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board_columns.RequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/dbms/v1/board_columns/{id}": {
            "get": {
                "description": "Get board column by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Get board column by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board column ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                }
            },
            "put": {
                "description": "Update board column",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Update board column",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board column ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update board column request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board_columns_dtos.BoardColumnsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete board column",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Delete board column",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board column ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/dbms/v1/board_columns/{id}/{field}": {
            "get": {
                "description": "Get board column field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Get board column field",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board column ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Field",
                        "name": "field",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                }
            },
            "put": {
                "description": "Update board column field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Update board column field",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board column ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Field",
                        "name": "field",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update board column request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwBoardColumn"
                        }
                    }
                }
            }
        },
...
        "/dbms/v1/workspace/{workspace_id}/board_columns": {
            "get": {
                "description": "Get board columns by workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board_columns"
                ],
                "summary": "Get board columns by workspace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TwBoardColumn"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "board_columns.RageRequest": {
            "type": "object",
            "properties": {
                "position1": {
                    "type": "integer"
                },
                "position2": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "board_columns.RequestBody": {
            "type": "object",
            "properties": {
                "position": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "board_columns_dtos.BoardColumnsRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwBoardColumn": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/models.TwWorkspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwWorkspace": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "extra_data": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}